requirements.txt
python-dotenv
yt-dlp
